feat: implement strategy-based context compaction with sliding window and summarization strategies
