+++ crates/forge_domain/src/compaction/summarization.rs
@@ -1,3 +1,4 @@
+use std::cmp::min;
 /// Summarization strategy for context compaction
 ///
 /// This strategy identifies sequences of assistant messages in a conversation
@@ -8,10 +9,11 @@ use std::sync::Arc;
 use anyhow::Result;
 use tracing::debug;
 
+use super::adjust_range::adjust_range_for_tool_calls;
 use super::strategy::CompactionImpact;
 use super::CompactionStrategy;
 use crate::services::{Services, TemplateService};
-use crate::{Compact, Context, ContextMessage};
+use crate::{Compact, Context, ContextMessage, Role};
 
 /// Compaction strategy that identifies sequences of messages and replaces them
 /// with a summary generated by an LLM
@@ -33,7 +35,7 @@ impl<S: Services> CompactionStrategy for SummarizationStrategy<S> {
 
     fn is_applicable(&self, compact: &Compact, context: &Context) -> bool {
         let preserve_last_n = compact.retention_window;
-        super::find_sequence(context, preserve_last_n).is_some()
+        find_sequence(context, preserve_last_n).is_some()
     }
 
     async fn compact(
@@ -45,7 +47,7 @@ impl<S: Services> CompactionStrategy for SummarizationStrategy<S> {
         let original_message_count = context.messages.len();
 
         // Find a sequence of messages to summarize
-        if let Some((start, end)) = super::find_sequence(&context, preserve_last_n) {
+        if let Some((start, end)) = find_sequence(&context, preserve_last_n) {
             debug!(
                 strategy = self.id(),
                 start, end, "Found compressible sequence of messages"
@@ -78,4 +80,79 @@ impl<S: Services> CompactionStrategy for SummarizationStrategy<S> {
             Ok((context, impact))
         }
     }
+}
+
+/// Finds all valid compressible sequences in the context, respecting the
+/// preservation window
+///
+/// This function identifies sequences of assistant messages between user
+/// messages that can be compressed by summarization, while respecting the
+/// preservation window.
+fn find_sequence(context: &Context, preserve_last_n: usize) -> Option<(usize, usize)> {
+    let messages = &context.messages;
+    if messages.is_empty() {
+        return None;
+    }
+
+    // len will be always > 0
+    let length = messages.len();
+    let mut max_len = length - min(length, preserve_last_n);
+
+    if max_len == 0 {
+        return None;
+    }
+
+    // Additional check: if max_len < 1, we can't safely do max_len - 1
+    if max_len < 1 {
+        return None;
+    }
+    if messages
+        .get(max_len - 1)
+        .is_some_and(|msg| msg.has_tool_call())
+    {
+        max_len -= 1;
+    }
+
+    let user_messages = messages
+        .iter()
+        .enumerate()
+        .take(max_len)
+        .filter(|(_, message)| message.has_role(Role::User))
+        .collect::<Vec<_>>();
+
+    // If there are no user messages, there can't be any sequences
+    if user_messages.is_empty() {
+        return None;
+    }
+    let start_positions = user_messages
+        .iter()
+        .map(|(start, _)| min(start.saturating_add(1), max_len.saturating_sub(1)))
+        .collect::<Vec<_>>();
+
+    let mut end_positions = user_messages
+        .iter()
+        .skip(1)
+        .map(|(pos, _)| pos.saturating_sub(1))
+        .collect::<Vec<_>>();
+    end_positions.push(max_len - 1);
+
+    // If either vector is empty, there can't be any compressible sequences
+    if start_positions.is_empty() || end_positions.is_empty() {
+        return None;
+    }
+
+    // Find a valid sequence and adjust it to preserve tool call chains
+    let range = start_positions
+        .iter()
+        .zip(end_positions.iter())
+        .find(|(start, end)| *end > *start)
+        .map(|(a, b)| (*a, *b));
+
+    // If we found a range, adjust it to respect tool call chains
+    if let Some((start, end)) = range {
+        let (adjusted_start, adjusted_end) = adjust_range_for_tool_calls(context, start, end);
+        return Some((adjusted_start, adjusted_end));
+    }
+
+    None
 }
